Fix the MD debug info to the `expand' kernels
As follow up of the change of [1] in llvm upstream, function metadata
should now be set through the function object. With [2], the above change
from upstream has been pulled into the LLVM version of the AOSP, requiring
to update RSAddDebugInfoPass.cpp to adapt to the new introduced logic.

Without this fix, the debug information for the 'kernel.expand'
functions is corrupt, making not possible to inspect the local variables
through a debugger.

References
[1] http://reviews.llvm.org/D14265
[2] https://android.googlesource.com/platform/frameworks/compile/libbcc/+/7a611ce47aed312cd92d2749dd4243d4384d2dd9

Change-Id: I95acb22d9d9fd0e2c28b859b01e84ed4a773e17b