bcc vendor plugin must be loaded from the sphal namespace
bcc loads vendor plugin in its address space. Since bcc is a system
process, the vendor plugin must be loaded from the sphal namespace where
VNDK-SP libraries are provided. This is to ensure that the vendor plugin
is linked against to the VNDK-SP version of libraries. It also prevents
the vendor plugin from using non-VNDK-SP libs such as libLLVM.so.

Bug: 62517583
Test: a sample renderscript app (CameraScript) runs well
Change-Id: I87c9d55c3527bee8c6d2653fa8a796143a7bb146